% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mi.R
\name{compute_mi}
\alias{compute_mi}
\title{Compute mutual information matrix}
\usage{
compute_mi(data)
}
\arguments{
\item{data}{Data matrix. This is an NxM matrix, where genes are on the rows and samples on the columns}
}
\value{
Returns the mutual information matrix for the data.
}
\description{
This function will use the `buid.mim` function in the `minet` package to compute the mutal information matrix in the gene space.
}
\examples{
M <- nsmblR::data_matrix
compute_mi(M)
}
