% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{Consensus network}
\usage{
consensus(vote_tally)
}
\arguments{
\item{vote_tally}{Data frame of tallied votes as defined in the \link{\code{edge_voting}} function}
}
\value{
A  data frame, filtered by super majority vote decision.
}
\description{
Based on the tallied votes using the \link{\code{edge_voting}} function, it provides the final consensus network.
}
