% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_matrix}
\alias{data_matrix}
\title{Example gene expression data
RNA-seq data from E. coli}
\format{An object of class \code{matrix} with 200 rows and 20 columns.}
\usage{
data_matrix
}
\description{
Example gene expression data
RNA-seq data from E. coli
}
\keyword{datasets}
