% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_filtering.R
\name{edge_filtering}
\alias{edge_filtering}
\title{Edge filtering}
\usage{
edge_filtering(inferred_network, quantile_thr)
}
\arguments{
\item{inferred_network}{An inferred networks as defined in the \code{\link{infer_network}} function}

\item{quantile_thr}{Quantile based threshold to threshold inferred networks. Defaults to the top 10% set of edges.}
}
\value{
A filtered data frame with edges being quantified (1) or not (0) based on the quantiles for each of the methods.
}
\description{
This function performs a filtering of the inference results based on a quantile calculation of the results. This is currently hard-coded so that the function selects the top 10% edges in each of the inference algorithms.
}
