% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Example gene names (E. coli)}
\format{An object of class \code{character} of length 200.}
\usage{
genes
}
\description{
Example gene names (E. coli)
}
\keyword{datasets}
