% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{infer_network}
\alias{infer_network}
\title{Infer network}
\usage{
infer_network(method = c(1, 2, 3, 4, 5, 6, 7), data, quantile_thr)
}
\arguments{
\item{method}{Integer defining the method to be used. Defaults to using CLR (method 3).}

\item{data}{Gene expression data frame or mutual information matrix}

\item{quantile_thr}{Quantile threshold for edge assignment (defaults to 0.95)}
}
\value{
An inferred network matrix
}
\description{
This function infers a gene expression network based on pre-defined methods.
}
\details{
Inference methods:
\enumerate{
  \item Spearman correlations
  \item Partial correlations with information theory (PCIT)
  \item Context Likelihood of Relatedness (CLR)
  \item ARACNe
  \item MRNET
  \item MRNETB
  \item Mutual Rank (MutRank)
}
}
