% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results.R
\name{merge_results}
\alias{merge_results}
\title{Merge results}
\usage{
merge_results(network_list)
}
\arguments{
\item{network_list}{A list of inferred networks as defined in the \code{\link{infer_network}} function.}
}
\value{
A data frame with all the results for the inference models used.
}
\description{
This function binds all the results from the inference algorithms into a single tidy data frame.
}
