% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsmblR.R
\docType{package}
\name{nsmblR}
\alias{nsmblR}
\alias{nsmblR-package}
\title{nsmblR: Network inference ensembler}
\description{
`nsmblR` is a package to perform inference of gene regulatory networks using an ensemble approach. `nsmblR` will take a gene expression compendium and infer the gene regulatory architecture using 7 popular network inference algorithms and then performing an edge count to find the consensus edges based on diverse voting methods.
}
