% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_filtering.R
\name{regulatory_filtering}
\alias{regulatory_filtering}
\title{Regulators filtering}
\usage{
regulatory_filtering(ensemble_df, organism)
}
\arguments{
\item{ensemble_df}{A data frame from \link{\code{merge_results}}}
}
\value{
A filtered data frame
}
\description{
This function will filter the inference results to only those between transcriptional regulators and protein coding genes. It assumes that gene-gene interactions happen due to the effect of a transcriptinal regulator. This function should be run independent of the outcome of edge filtering done with the \link{\code{edge_filtering}} function.
}
