% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttrust_data}
\alias{ttrust_data}
\title{TTRUST data
Data on TF -> Gene interactions in human annotated by TTRUST}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9396 rows and 4 columns.}
\usage{
ttrust_data
}
\description{
TTRUST data
Data on TF -> Gene interactions in human annotated by TTRUST
}
\keyword{datasets}
