% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_voting.R
\name{edge_voting}
\alias{edge_voting}
\title{Edge voting}
\usage{
edge_voting(ensemble_df)
}
\arguments{
\item{ensemble_df}{A data frame with the outcomes of all inference methods, as defined in the \code{\link{merge_results}} or in \code{\link{regulatory_filtering}} functions}
}
\value{
A data frame with the outcome of chosen voting method
}
\description{
Method to define ensemble method. Given a data frame containing inferred relationships between genes, the method will apply a defined voting method to define  a consensus network.
Voting methods:
}
