% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_voting.R
\name{edge_voting}
\alias{edge_voting}
\title{Edge voting}
\usage{
edge_voting(ensemble_df)
}
\arguments{
\item{ensemble_df}{A data frame with the outcomes of all inference methods, as defined in the \code{\link{merge_results}} or in \code{\link{regulatory_filtering}} functions}
}
\value{
A data frame with the outcome of chosen voting method
}
\description{
Method to define ensemble method. Given a data frame containing inferred relationships between genes, the method will apply a defined voting method to define  a consensus network. Note that the function will perform all of these voting procedures on the data.
}
\details{
Voting methods:
\enumerate{
  \item Majority vote: edge is considered present if it is inferred in 51% of the cases
  \item Super majority vote: edge is considered present if it is inferred in 2/3 of the cases
  \item Quorum vote: this method sits between between the majority vote and the supermajority. It is defined as \eqn{\frac{N}{2} + 1}, where N is the number of methods used
  \item Absolute majority vote: edge is present in \strong{all} of the methods only
}
}
