% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_model.R
\name{ensemble_model}
\alias{ensemble_model}
\title{Ensemble model (wrapper for nsmblR)}
\usage{
ensemble_model(data, gene_names)
}
\arguments{
\item{data}{Gene expression data. Matrix is NxM, with genes on the rows and samples on the column.}

\item{gene_names}{Names for the genes in the data set in the same order as the expression data matrix.}
}
\value{
Returns a list containing the inferred networks, the filtered networks (based on regulatory interactions) and a ranked data frame in accordance to the \code{\link{consensus}} function.
}
\description{
This function will generate an ensemble gene regulatory network model based on different inference algorithms.
}
