% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{infer_network}
\alias{infer_network}
\title{Infer network}
\usage{
infer_network(method = c(1, 2, 3, 4, 5, 6, 7), data)
}
\arguments{
\item{method}{Integer defining the method to be used. Defaults to using CLR (method 3).}

\item{data}{Gene expression data frame or mutual information matrix}
}
\value{
An inferred network matrix
}
\description{
This function infers a gene expression network based on pre-defined methods.
}
\details{
Inference methods:
\itemize{
  \item{1}{Spearman correlations}
  \item{2}{Partial correlations with information theory (PCIT)}
  \item{3}{Context Likelihood of Relatedness (CLR)}
  \item{4}{ARACNe}
  \item{5}{MRNET}
  \item{6}{MRNETB}
  \item{7}{Mutual Rank (MutRank)}
}
}
