% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_network.R
\name{infer_network}
\alias{infer_network}
\title{Infer network}
\usage{
infer_network(method = c(1, 2, 3, 4, 5, 6, 7), data)
}
\arguments{
\item{method}{Integer defining the method to be used}

\item{data}{Gene expression data frame or mutual information matrix}
}
\value{
An inferred network matrix
}
\description{
This function infers a gene expression network based on pre-defined methods. These are:
1. Spearman correlations
2. Partial correlations with information theory (PCIT)
3. Context Likelihood of Relatedness (CLR)
4. ARACNe
5. MRNET
6. MRNETB
7. Mutual Rank (MutRank)
}
