% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_filtering.R
\name{regulatory_filtering}
\alias{regulatory_filtering}
\title{Regulators filtering}
\usage{
regulatory_filtering(ensemble_df, organism, gene_names)
}
\arguments{
\item{ensemble_df}{A data frame from \code{\link{merge_results}}}

\item{organism}{Organism being analyzed (E. coli or Human)}

\item{gene_names}{Set of gene names in the data set (ordered as in the expression data matrix)}
}
\value{
A filtered data frame
}
\description{
This function will filter the inference results to only those between transcriptional regulators and protein coding genes. It assumes that gene-gene interactions happen due to the effect of a transcriptinal regulator. This function should be run independent of the outcome of edge filtering done with the \code{\link{edge_filtering}} function.
}
\examples{
regulatory_filtering(N, 1, nsmblR::genes)
}
